+++ package.json
@@ -1,43 +1,77 @@
 {
-  "name": "metacoders-courses-service",
+  "name": "nest-typescript-starter",
+  "private": true,
   "version": "1.0.0",
-  "description": "",
-  "main": "dist/index.js",
+  "description": "Nest TypeScript starter repository",
+  "license": "MIT",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development concurrently \"npx tsc --watch\" \"nodemon -q dist/src/index.js\"",
-    "build": "rimraf dist && npx tsc",
-    "start": "cross-env NODE_ENV=production concurrently \"npm run build\"  \"node dist/src/index.js\"",
-    "prisma:generate": "ts-node prisma/seed.ts"
+    "prebuild": "rimraf dist",
+    "build": "nest build",
+    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
+    "start": "nest start",
+    "start:dev": "nest start --watch",
+    "start:debug": "nest start --debug --watch",
+    "start:prod": "node dist/main",
+    "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:cov": "jest --coverage",
+    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
+    "test:e2e": "jest --config ./test/jest-e2e.json"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "dependencies": {
-    "@prisma/client": "^4.2.1",
-    "@types/chance": "^1.1.3",
+    "@nestjs/common": "^9.0.0",
+    "@nestjs/core": "^9.0.0",
+    "@nestjs/platform-express": "^9.0.0",
+    "@prisma/client": "^4.3.1",
+    "@types/cheerio": "^0.22.31",
     "@types/lodash": "^4.14.184",
     "axios": "^0.27.2",
-    "body-parser": "^1.20.0",
-    "chance": "^1.1.8",
     "cheerio": "^1.0.0-rc.12",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.1",
-    "express": "^4.18.1",
     "lodash": "^4.17.21",
-    "nanoid": "^4.0.0",
-    "puppeteer": "^17.0.0",
-    "rimraf": "^3.0.2"
+    "reflect-metadata": "^0.1.13",
+    "rimraf": "^3.0.2",
+    "rxjs": "^7.5.5"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.12",
+    "@nestjs/cli": "^9.0.0",
+    "@nestjs/schematics": "^9.0.0",
+    "@nestjs/testing": "^9.0.0",
     "@types/express": "^4.17.13",
-    "@types/node": "^18.7.3",
-    "@types/puppeteer": "^5.4.6",
-    "concurrently": "^7.3.0",
-    "cross-env": "^7.0.3",
-    "nodemon": "^2.0.19",
-    "prisma": "^4.2.1",
-    "ts-node": "^10.9.1",
+    "@types/jest": "^28.1.4",
+    "@types/node": "^18.0.3",
+    "@types/supertest": "^2.0.12",
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
+    "eslint": "^8.19.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^28.1.2",
+    "prettier": "^2.7.1",
+    "prisma": "^4.3.1",
+    "source-map-support": "^0.5.21",
+    "supertest": "^6.2.4",
+    "ts-jest": "^28.0.5",
+    "ts-loader": "^9.3.1",
+    "ts-node": "^10.8.2",
+    "tsconfig-paths": "^4.0.0",
     "typescript": "^4.7.4"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "ts"
+    ],
+    "rootDir": "src",
+    "testRegex": ".*\\.spec\\.ts$",
+    "transform": {
+      "^.+\\.(t|j)s$": "ts-jest"
+    },
+    "collectCoverageFrom": [
+      "**/*.(t|j)s"
+    ],
+    "coverageDirectory": "../coverage",
+    "testEnvironment": "node"
   }
 }
