+++ src/dataMockup/dataMockup.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+import { DataMockupService } from "./dataMockup.service";
+
+@Module({
+  exports: [DataMockupService],
+  providers: [DataMockupService]
+})
+
+export class DataMockupModule {}
