--- src/interfaces/courseGetInterface/CourseGetResponse.ts
@@ -1,14 +1,0 @@
-export interface CourseGetResponse {
-  id: string;
-  title: string;
-  imageUrl: string;
-  videoUrl: string;
-  description?: string;
-  shortDescription?: string;
-  vote?: number;
-  during?: string;
-  createdAt: string;
-  updatedAt: string;
-  promoCode?: string;
-  price?: number;
-};
