--- src/interfaces/topicGetResponse/TopicGetResponse.ts
@@ -1,7 +1,0 @@
-export interface TopicGetResponse {
-  id: string,
-  title: string,
-  description: string,
-  createdAt: string,
-  updatedAt: string,
-};
