--- src/services/counts/countsCourse/countsCourseInteractor.ts
@@ -1,5 +1,0 @@
-import { Course } from "../../../dataMockup/DataMockupInterface";
-
-export const countsCourseInteractor = (data: Course[]): number => {
-  return data.length;
-}
