+++ src/services/counts/countsCourses/countsCourses.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { DataMockupService } from "../../../dataMockup/dataMockup.service";
+
+@Injectable()
+export class CountsCoursesService {
+  constructor(private dataMockupService: DataMockupService) {}
+
+  execute(): number {
+    const courses = this.dataMockupService.excute();
+    return courses.length;
+  }
+}
