+++ src/services/courses/common/coursesGetByOffset.service.ts
@@ -1,0 +1,17 @@
+import { Injectable } from "@nestjs/common";
+import { Course } from "../../../dataMockup/DataMockupInterface";
+// import { chunk } from "../../../pkgs/lodash";
+import chunk from "lodash/chunk";
+
+@Injectable()
+export class CoursesGetByOffsetService {
+
+  execute(data: Course[], limit: number, offset: number): Course[] {
+    if (offset === 1) {
+      const courses = chunk(data, limit)[0];
+      return courses || data;
+    }
+    const courses = chunk(data, (offset - 1) * limit)[1];
+    return courses || [];
+  }
+}
