+++ test/app.e2e-spec.ts
@@ -1,0 +1,24 @@
+import * as request from 'supertest';
+import { Test } from '@nestjs/testing';
+import { AppModule } from './../src/app.module';
+import { INestApplication } from '@nestjs/common';
+
+describe('AppController (e2e)', () => {
+  let app: INestApplication;
+
+  beforeAll(async () => {
+    const moduleFixture = await Test.createTestingModule({
+      imports: [AppModule],
+    }).compile();
+
+    app = moduleFixture.createNestApplication();
+    await app.init();
+  });
+
+  it('/ (GET)', () => {
+    return request(app.getHttpServer())
+      .get('/')
+      .expect(200)
+      .expect('Hello World!');
+  });
+});
